Add a few aliases and use new API in interp functions
