+++ src/lang.c
@@ -1,5 +1,4 @@
-#define R_NO_REMAP
-#include <Rinternals.h>
+#include "rlang.h"
 
 SEXP r_new_language_(SEXP head, SEXP tail) {
   return Rf_lcons(head, tail);
@@ -10,4 +9,8 @@ SEXP r_new_language(SEXP head, SEXP tail) {
   SEXP out = Rf_lcons(head, tail);
   UNPROTECT(2);
   return out;
+}
+
+bool r_is_lang(SEXP x) {
+  return TYPEOF(x) == LANGSXP;
 }
