+++ src/lang.h
@@ -5,5 +5,7 @@
 SEXP r_new_language_(SEXP head, SEXP tail);
 SEXP r_new_language(SEXP head, SEXP tail);
 
+bool r_is_lang(SEXP x);
+
 
 #endif
